Execute

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,6 +1,7 @@
 # Pkill
 exec {'killmenow':
 command  => 'pkill -f killmenow',
-provider => 'exec',
+provider => shell,
+onlyif => '/usr/bin/ -e /usr/bin',
 }
